pyproject.toml
[project]
name = "agentic-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.120.0",
    "jupyter>=1.1.1",
    "matplotlib>=3.8.0",
    "motor>=3.7.1",
    "openai>=2.3.0",
    "pydantic>=2.12.0",
    "pymongo>=4.15.3",
    "pypdf>=6.1.3",
    "python-dotenv>=1.1.1",
    "seaborn>=0.13.0",
    "uvicorn>=0.38.0",
]
